suggestions_api/requirements.txt
elasticsearch==7.17.0
fastapi==0.103.2
httpx==0.25.0
pandas==2.1.1
pydantic==2.4.2
pytest==7.4.2
pytest-cov==4.1.0
python-dotenv==1.0.0
uvicorn==0.23.2